Added android posts screen

+++ shared/src/commonMain/kotlin/com/example/kmmtutorial/models/Post.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Post(
     val body: String,
-    val id: Int,
+    val id: Long,
     val title: String,
-    val userId: Int
+    val userId: Long
 )
